requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
youtube-transcript-api
Pillow
